Starting off point
Third version from the beginning, and the first that can be used as a good starting-off point.

Current functionality:
MessageNode that gives a description of the situation and two options for the scenario. Ties to two MessageOutcomes and performs this outcome by reading user input and redirecting to one of these outcomes. The outcomes need to be tied to the node outside of their initial declaration.

MessageOutcome, which spits out the result of the choice and points to a new MessageNode. Is used as a parent for more complex outcomes, but also used for simple outcomes

RandomMessageOutcome, who has a random chance of giving a different result and pointing to a different node.

AddItemOutcome, which adds an object of type Item to the Inventory

NeedItemOutcome, which checks the Inventory for the specified item and only allows progress if the player has it

RandomItemOutcome, which acts as random message, but if randomness happens it also adds an Item.

Item, mostly used for checks of progress. It has a string component for its name

Inventory, which keeps track of all Items. Is turned into a singleton, can have Items added to it, specific Items removed from it, checked to spit out a list of its current items (currently not callable from in-game)